Move reduction rule into policy
